Added Initial interface for program
This file has the capability to open a pop up that will write to the logfile. Any writes to the logfile will come with a timestamp that includes exact time and date.
It also has an option for offload selection where the user can upload a manifest and select the containers that need to be removed. The container info is then stored
into an array and send to the calculate function to find optimal path