Added Animation for moves to be performed
The animation will create a grid based off the manifest. Then it will receieve a 2D array of paths which is just an array that holds the coordinates for the optimal path of the move.
Right now it is using a sample 2D array.
TODO: integrate this functionality with the main interface
TODO: Change to receive 2D array of paths from calculate function