Updated animation with better visuals
There is now a 9th row which sits above the ship because when onloading and offloading it may be used. Also made the animation account for containers on the ship so they now appear green. If a container is moved to a difference place on the ship, the original spot becomes white again while the new spot beomces green. It adds/removes containers to account for onload and offload.