Program logfile now accounts for crashes
The program now reads to the old log file instead of a hard coded one. Example, logfiles.txt holds all logfiles and now if user does not want to create a new logfile, it will read the last logfile in logfiles.txt to write to. This accounts for the program crashing and not needing to create a new logfile.